Auth

+++ controllers/cards.js
@@ -1,54 +1,61 @@
 const mongoose = require('mongoose');
-const { codesError } = require('../const');
 const Card = require('../models/card');
+const IncorrectError = require('../errors/incorrectError');
+const ForbiddenError = require('../errors/forbiddenError');
+const NotFoundDataError = require('../errors/notFoundDataError');
 
-const getCards = (req, res) => {
+const getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => {
-      res.status(codesError.DEFAULT).send({ message: 'На сервере произошла ошибка' });
-    });
+    .catch(next);
 };
 
-const postCard = (req, res) => {
+const postCard = (req, res, next) => {
   const { name, link } = req.body;
 
   Card.create({ name, link, owner: req.user._id })
     .then((card) => res.status(201).send({ data: card }))
     .catch((err) => {
       if (err instanceof mongoose.Error.ValidationError) {
-        res.status(codesError.INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные при создании карточки',
-        });
+        next(
+          new IncorrectError(
+            'Переданы некорректные данные при создании карточки'
+          )
+        );
       } else {
-        res.status(codesError.DEFAULT).send({ message: 'На сервере произошла ошибка' });
+        next(err);
       }
     });
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
+const deleteCard = (req, res, next) => {
+  Card.findById(req.params.cardId)
     .then((card) => {
       if (card) {
-        res.send({ data: card });
+        const owner = card.owner.toString();
+
+        if (owner !== req.user._id) {
+          throw new ForbiddenError('У вас нет доступа');
+        } else {
+          return Card.findByIdAndRemove(req.params.cardId);
+        }
       } else {
-        res
-          .status(codesError.NOT_FOUND_DATA)
-          .send({ message: 'Карточка с указанным id не найдена' });
+        throw new NotFoundDataError('Карточка с указанным id не найдена');
       }
+    })
+    .then((card) => {
+      res.send({ data: card });
     })
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
-        res.status(codesError.INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные',
-        });
+        next(new IncorrectError('Переданы некорректные данные'));
       } else {
-        res.status(codesError.DEFAULT).send({ message: 'На сервере произошла ошибка' });
+        next(err);
       }
     });
 };
 
-const putLikeCard = (req, res) => {
+const putLikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $addToSet: { likes: req.user._id } },
@@ -58,23 +65,23 @@ const putLikeCard = (req, res) => {
       if (card) {
         res.send({ data: card });
       } else {
-        res
-          .status(codesError.NOT_FOUND_DATA)
-          .send({ message: 'Передан несуществующий id карточки' });
+        throw new NotFoundDataError('Передан несуществующий id карточки');
       }
     })
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
-        res.status(codesError.INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные для установки лайка',
-        });
+        next(
+          new IncorrectError(
+            'Переданы некорректные данные для установки лайка'
+          )
+        );
       } else {
-        res.status(codesError.DEFAULT).send({ message: 'На сервере произошла ошибка' });
+        next(err);
       }
     });
 };
 
-const deleteLikeCard = (req, res) => {
+const deleteLikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(
     req.params.cardId,
     { $pull: { likes: req.user._id } },
@@ -84,18 +91,16 @@ const deleteLikeCard = (req, res) => {
       if (card) {
         res.send({ data: card });
       } else {
-        res
-          .status(codesError.NOT_FOUND_DATA)
-          .send({ message: 'Передан несуществующий id карточки' });
+        throw new NotFoundDataError('Передан несуществующий id карточки');
       }
     })
     .catch((err) => {
       if (err instanceof mongoose.Error.CastError) {
-        res.status(codesError.INCORRECT_DATA).send({
-          message: 'Переданы некорректные данные для удаления лайка',
-        });
+        next(
+          new IncorrectError('Переданы некорректные данные для удаления лайка')
+        );
       } else {
-        res.status(codesError.DEFAULT).send({ message: 'На сервере произошла ошибка' });
+        next(err);
       }
     });
 };
